Move new code to individual classes
