+++ Core/Runtime/Kegstand/Simulator.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using JetBrains.Annotations;
 using Kegstand.Impl;
 using UnityEngine.Assertions;
@@ -16,102 +15,11 @@ namespace Kegstand
         void Update(float deltaTime);
     }
 
-    public interface TimedEventQueue : IList<TimedEvent>
-    {
-        TimedEvent EnqueueNewEventToBuffer(Keg keg, float deltaTimeUnit, KegEvent eventType);
-    }
-
-    [Obsolete("Temporarily created for refactoring. Will be updated after preparing unit testing")]
-    public class TempTimedEventQueue<TTimeValue> : List<TimedEvent<TTimeValue>>, TimedEventQueue 
-        where TTimeValue : IComparable<TTimeValue>
-    {
-        [NotNull]
-        private readonly Func<float, TTimeValue> timeUnitConverter;
-
-        public TempTimedEventQueue(Func<float, TTimeValue> timeUnitConverter)
-        {
-            Assert.IsNotNull(timeUnitConverter);
-            this.timeUnitConverter = timeUnitConverter;
-        }
-
-        #region IList<TimedEvent>
-        IEnumerator<TimedEvent> IEnumerable<TimedEvent>.GetEnumerator()
-        {
-            return this.Cast<TimedEvent>().GetEnumerator();
-        }
-        
-        void ICollection<TimedEvent>.Add(TimedEvent item)
-        {
-            if (item is TimedEvent<TTimeValue> typedItem)
-            {
-                Add(typedItem);
-            }
-        }
-
-        bool ICollection<TimedEvent>.Contains(TimedEvent item)
-        {
-            if (item is TimedEvent<TTimeValue> typedItem)
-            {
-                return Contains(typedItem);
-            }
-            return false;
-        }
-
-        public void CopyTo(TimedEvent[] array, int arrayIndex)
-        {
-            throw new NotImplementedException();
-        }
-
-        bool ICollection<TimedEvent>.Remove(TimedEvent item)
-        {
-            if (item is TimedEvent<TTimeValue> typedItem)
-            {
-                return Remove(typedItem);
-            }
-            return false;
-        }
-
-        bool ICollection<TimedEvent>.IsReadOnly => false;
-        
-        int IList<TimedEvent>.IndexOf(TimedEvent item)
-        {
-            throw new NotImplementedException();
-        }
-
-        void IList<TimedEvent>.Insert(int index, TimedEvent item)
-        {
-            throw new NotImplementedException();
-        }
-
-        TimedEvent IList<TimedEvent>.this[int index]
-        {
-            get => throw new NotImplementedException();
-            set => throw new NotImplementedException();
-        }
-        #endregion  IList<TimedEvent>
-
-        public TimedEvent EnqueueNewEventToBuffer(Keg keg, float deltaTimeUnit, KegEvent eventType)
-        {
-            var timedEvent = new TimedEvent<TTimeValue>(keg, timeUnitConverter(deltaTimeUnit), eventType);
-            Add(timedEvent);
-            return timedEvent;
-        }
-    }
-
-    public class SimulatorFactory
-    {
-        public static Simulator<TimeSpan, TimeSpanClock> CreateDefault()
-        {
-            var amountVisitor = new TimeSpanAmountVisitor(new Timestamp<TimeSpan>(TimeSpan.Zero));
-            return new Simulator<TimeSpan, TimeSpanClock>(amountVisitor, timeUnit=> TimeSpan.FromSeconds(timeUnit));
-        }
-    }
-    
     public class Simulator<TTimeValue, TClock> : Simulator 
         where TClock : class, Clock<TTimeValue>, new()
         where TTimeValue : IComparable<TTimeValue>
     {
-        [NotNull] private readonly TempTimedEventQueue<TTimeValue> timedEventsScratchList;
+        [NotNull] private readonly TimedEventQueue<TTimeValue> timedEventsScratchList;
         [NotNull] private AmountVisitor<TTimeValue> amountVisitor;
 
         public readonly IReadOnlyList<TimedEvent<TTimeValue>> Events;
@@ -126,11 +34,12 @@ namespace Kegstand
         [NotNull]
         private readonly TClock clock;
 
-        public Simulator(AmountVisitor<TTimeValue> amountVisitor, Func<float, TTimeValue> unitTimeConverter)
+        public Simulator(TimedEventQueue<TTimeValue> eventQueue, AmountVisitor<TTimeValue> amountVisitor)
         {
+            Assert.IsNotNull(eventQueue);
             Assert.IsNotNull(amountVisitor);
-            
-            timedEventsScratchList = new TempTimedEventQueue<TTimeValue>(unitTimeConverter);
+
+            timedEventsScratchList = eventQueue;
             this.amountVisitor = amountVisitor;
             
             clock = new TClock();
