+++ Core/Runtime/Kegstand/TimedEventQueue.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace Kegstand
+{
+    public interface TimedEventQueue : IList<TimedEvent>
+    {
+        TimedEvent EnqueueNewEventToBuffer(Keg keg, float deltaTimeUnit, KegEvent eventType);
+    }
+}
